+++ CHANGELOG.md
@@ -12,6 +12,10 @@ I'm new at this, so expect imperfection ;_; I'm trying!
 ### Added
 
 + Added `--all` flag that displays all games
++ Added long help messages invoked with `--help`
++ Added some short argument invocations
+  + `-r` for `--archive-root`
+  + `-s` for `--systems`
 
 ### Changed